Attendes

+++ src/events/events.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Delete, Get, Post, Patch, Param, Body, HttpCode, ParseIntPipe, ValidationPipe, UsePipes } from '@nestjs/common';
+import { Controller, Delete, Get, Post, Patch, Param, Body, HttpCode, ParseIntPipe,  UsePipes, Logger, HttpException, NotFoundException } from '@nestjs/common';
 import { CreateEventDto } from './create-event.dto';
 import { UpdateEventDto } from './update-event.dto';
 import { Event } from './event.entity';
@@ -7,24 +7,40 @@ import { InjectRepository } from '@nestjs/typeorm';
 
 @Controller('/events')
 export class EventsController {
+  private readonly logger = new Logger(EventsController.name);
 
   constructor(
     @InjectRepository(Event)
     private readonly repository:Repository<Event>
   ) { }
 
-  private events:Event[] = [];
-
 
   @Get()
   async findAll(){
-    return await this.repository.find();
+    try {
+      this.logger.log(`Finding all event`);
+      return await this.repository.find();  
+    } 
+    catch (error) {
+
+        this.logger.log(error);
+        throw new HttpException(`Internal Server Error ${error.message}`, 500);
+      }    
   }
 
   @Get(':id')
   async findOne(@Param('id', ParseIntPipe) id){
+    
+    const event = await this.repository.find({
+      where: {
+        id
+      }
+    });
 
-    const event = await this.repository.findOne(id);
+    if(!event){
+      throw new NotFoundException(`Event with id ${id} not found`);
+    }
+
     return event;
 
   }
@@ -33,7 +49,7 @@ export class EventsController {
   @Post()
   @HttpCode(201)
   async create(@Body() input:CreateEventDto) {
-
+    
     const event={
       ... input,
       when: new Date(input.when)
@@ -50,6 +66,10 @@ export class EventsController {
     
     const event =await this.repository.findOne(id);
 
+    if(!event){
+      throw new NotFoundException(`Event with id ${id} not found`);
+    }
+
     return await this.repository.save(   {
       ... event,
       ... input,
@@ -63,6 +83,10 @@ export class EventsController {
   async remove(@Param('id') id) {
 
     const event=await this.repository.findOne(id);
+
+    if(!event){
+      throw new NotFoundException(`Event with id ${id} not found`);
+    }
 
     await this.repository.remove(event);
